最后修改

+++ login/src/main/java/com/zb/manage/biz/service/StudentService.java
@@ -1,7 +1,9 @@
 package com.zb.manage.biz.service;
 
+import com.zb.manage.dal.mapper.RecordMapper;
+import com.zb.manage.dal.mapper.SelectionMapper;
 import com.zb.manage.dal.mapper.StudentMapper;
-import com.zb.manage.dal.model.Student;
+import com.zb.manage.dal.model.StudentDO;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
@@ -14,16 +16,22 @@ public class StudentService {
     @Resource
     private StudentMapper studentMapper;
 
-    public List<Student> findAll() {
+    @Resource
+    private SelectionMapper selectionMapper;
+
+    @Resource
+    private RecordMapper recordMapper;
+
+    public List<StudentDO> findAll() {
         return studentMapper.findAll();
     }
 
-    public Student getByName(String name) {
+    public StudentDO getByName(String name) {
         return studentMapper.findByName(name);
     }
 
     public void add(BigInteger id, String name, BigInteger number, String age, String entime) {
-        Student student = new Student();
+        StudentDO student = new StudentDO();
         student.setId(id);
         student.setName(name);
         student.setAge(age);
@@ -34,18 +42,26 @@ public class StudentService {
     }
 
     public void delete(BigInteger id) {
+        int selectionCount = selectionMapper.countByStudentId(id);
+        if (selectionCount > 0) {
+            throw new IllegalArgumentException("学生表有依赖此课程，不得删除");
+        }
+        int recordCount = recordMapper.countByStudentId(id);
+        if(recordCount > 0) {
+            throw new IllegalArgumentException("学生表有依赖此课程，不得删除");
+        }
         studentMapper.delete(id);
     }
 
-    public void update(Student student) {
+    public void update(StudentDO student) {
         studentMapper.update(student);
     }
 
-    public Student getById(BigInteger id) {
+    public StudentDO getById(BigInteger id) {
         return studentMapper.getById(id);
     }
 
-    public Student getByAge(String age) {
+    public StudentDO getByAge(String age) {
         return studentMapper.findByAge(age);
     }
 }
